+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/affixes/adders/ManaArmorAffixes.java
@@ -9,7 +9,7 @@ import com.robertx22.age_of_exile.database.data.stats.types.resources.mana.Mana;
 import com.robertx22.age_of_exile.uncommon.enumclasses.ModType;
 import com.robertx22.library_of_exile.registry.ExileRegistryInit;
 
-public class MagicShieldPrefixes implements ExileRegistryInit {
+public class ManaArmorAffixes implements ExileRegistryInit {
 
     @Override
     public void registerAll() {
@@ -17,21 +17,21 @@ public class MagicShieldPrefixes implements ExileRegistryInit {
         AffixBuilder.Normal("glimmering")
             .Named("Glimmering")
             .stats(new StatModifier(5, 20, Mana.getInstance(), ModType.FLAT), new StatModifier(5, 15, Armor.getInstance(), ModType.FLAT))
-            .includesTags(SlotTag.magic_shield_stat)
+            .includesTags(SlotTag.armor_family)
             .Prefix()
             .Build();
 
         AffixBuilder.Normal("seraphim")
             .Named("Seraphim's")
             .stats(new StatModifier(5, 20, Mana.getInstance(), ModType.FLAT), new StatModifier(3, 10, Health.getInstance(), ModType.FLAT))
-            .includesTags(SlotTag.magic_shield_stat)
+            .includesTags(SlotTag.armor_family)
             .Prefix()
             .Build();
 
         AffixBuilder.Normal("azure")
             .Named("Azure")
             .stats(new StatModifier(5, 25, Mana.getInstance(), ModType.FLAT))
-            .includesTags(SlotTag.magic_shield_stat)
+            .includesTags(SlotTag.armor_family)
             .Prefix()
             .Build();