+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/base_gear_types/BaseGearsRegister.java
@@ -1,6 +1,9 @@
 package com.robertx22.age_of_exile.aoe_data.database.base_gear_types;
 
-import com.robertx22.age_of_exile.aoe_data.database.base_gear_types.adders.*;
+import com.robertx22.age_of_exile.aoe_data.database.base_gear_types.adders.BaseGearJewelry;
+import com.robertx22.age_of_exile.aoe_data.database.base_gear_types.adders.BaseGearOffHands;
+import com.robertx22.age_of_exile.aoe_data.database.base_gear_types.adders.BaseGearWeapons;
+import com.robertx22.age_of_exile.aoe_data.database.base_gear_types.adders.BasePlateArmors;
 import com.robertx22.library_of_exile.registry.ExileRegistryInit;
 
 public class BaseGearsRegister implements ExileRegistryInit {
@@ -8,8 +11,6 @@ public class BaseGearsRegister implements ExileRegistryInit {
     @Override
     public void registerAll() {
 
-        new BaseClothArmors().registerAll();
-        new BaseLeatherArmors().registerAll();
         new BasePlateArmors().registerAll();
 
         new BaseGearOffHands().registerAll();