+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/unique_gears/runewords/HelmetRunewords.java
@@ -1,6 +1,6 @@
 package com.robertx22.age_of_exile.aoe_data.database.unique_gears.runewords;
 
-import com.robertx22.age_of_exile.aoe_data.database.base_gear_types.adders.BaseClothArmors;
+import com.robertx22.age_of_exile.aoe_data.database.base_gear_types.adders.BasePlateArmors;
 import com.robertx22.age_of_exile.aoe_data.database.unique_gears.UniqueGearBuilder;
 import com.robertx22.age_of_exile.database.data.StatModifier;
 import com.robertx22.age_of_exile.database.data.stats.types.generated.ElementalResist;
@@ -22,7 +22,7 @@ public class HelmetRunewords implements ExileRegistryInit {
         UniqueGearBuilder.of(
                 "insight",
                 "Insight",
-                BaseClothArmors.HELMETS.get(LevelRanges.START_TO_END))
+                BasePlateArmors.HELMETS.get(LevelRanges.START_TO_END))
             .setReplacesName()
             .stats(Arrays.asList(
                 new StatModifier(20, 30, ManaRegen.getInstance(), ModType.PERCENT),