+++ src/main/java/com/robertx22/age_of_exile/uncommon/testing/TestManager.java
@@ -1,7 +1,5 @@
 package com.robertx22.age_of_exile.uncommon.testing;
 
-import com.robertx22.age_of_exile.uncommon.testing.tests.CheckWeaponDpsBalanceTest;
-import com.robertx22.age_of_exile.uncommon.testing.tests.CountUniqueGearTypes;
 import com.robertx22.age_of_exile.uncommon.testing.tests.MobTypesTest;
 import net.minecraft.server.world.ServerWorld;
 
@@ -10,8 +8,6 @@ public class TestManager {
     public static void RunAllTests(ServerWorld world) {
 
         if (false) {
-            CountUniqueGearTypes.count();
-            CheckWeaponDpsBalanceTest.run();
             MobTypesTest.run(world);
         }